% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_name_pattern.R
\name{assert_name_pattern}
\alias{assert_name_pattern}
\title{Ensure that all files in a vector are correctly named}
\usage{
assert_name_pattern(file, pattern, verbose = FALSE)
}
\arguments{
\item{file}{(vector of) file name(s)}

\item{pattern}{format for a study, generally in regular expression. See e.g. \url{https://regexr.com/} for help in creating suitable patterns}
}
\value{
vector of indices of files that don't pass the check
}
\description{
This function runs a list of file names through a pattern check. A warning
will be returned in case not all file names pass the check. Offending
file names will be shown.
}
\examples{
path <- "~/Documents/projects/toolmarks/data/x3p_files" # HH sorry for using absolute paths, Maria, you will need to exchange that with the path to the x3pfiles on your machine
files <- dir(path, pattern="x3p", recursive = TRUE, full.names=FALSE)
assert_name_pattern(files, pattern="^Sc\\\\d\\\\d\\\\.Pl\\\\d\\\\d\\\\d\\\\.Ma\\\\d\\\\.S[AB]\\\\.An\\\\d\\\\d\\\\.Pb\\\\.Dir(Fo|Ba)\\\\.\\\\Siz[SML](|.mid)\\\\.x3p$")
}
