% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_get_signature.R
\name{cc_get_signature}
\alias{cc_get_signature}
\title{Extract signature from crosscut}
\usage{
cc_get_signature(ccdata, span1 = 125, span2 = 5, resolution = 0.645)
}
\arguments{
\item{ccdata}{crosscut as returned from x3p_crosscut}

\item{span1}{The span for the loess fit to get from the profile to the raw
signature}

\item{span2}{The span for the loess fit to smooth the raw signature}

\item{resolution}{numeric value of  the scan resolution, see \code{x3ptools::x3p_get_scale}.}
}
\value{
data frame
}
\description{
x3p file of a 3d topological surface is processed at surface
crosscut y,
}
\examples{
if (interactive()) {
library(x3ptools)
x3p <- x3p_read("~/Documents/CSAFE/Wirecutter/data/Aluminum Wires renamed/T1AW-LI-R1-B01.x3p")
line1 <- x3p \%>\% x3p_extract_profile()
line1 <- line1$line_df
signature <- cc_get_signature(line1, resolution = x3p \%>\% x3p_get_scale())
}
}
