% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{Wrapper function for toolmarks data set in tmaRks package}
\usage{
data(
  ...,
  list = character(),
  package = NULL,
  lib.loc = NULL,
  verbose = getOption("verbose"),
  envir = .GlobalEnv,
  overwrite = TRUE
)
}
\arguments{
\item{...}{literal character strings or names.}

\item{list}{a character vector.}

\item{package}{
    a character vector giving the package(s) to look
    in for data sets, or \code{NULL}.

    By default, all packages in the search path are used, then
    the \file{data} subdirectory (if present) of the current working
    directory.
  }

\item{lib.loc}{a character vector of directory names of \R libraries,
    or \code{NULL}.  The default value of \code{NULL} corresponds to all
    libraries currently known.}

\item{verbose}{a logical.  If \code{TRUE}, additional diagnostics are
    printed.}

\item{envir}{the \link{environment} where the data should be loaded.}

\item{overwrite}{logical: should existing objects of the same name in
    \env{envir} be replaced?}
}
\description{
This function is specific to the\code{toolmarks} dataset.
It unpacks the information stored in the tool ID.
This allows us to store the data in a smaller dataset but unpack it for the
analysis.
}
\examples{
data(toolmarks)
dim(toolmarks)
}
