% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_align.R
\name{sig_align}
\alias{sig_align}
\title{Align two surface cross cuts according to maximal correlation}
\usage{
sig_align(
  sig1,
  sig2,
  min.overlap = round(0.75 * min(length(sig1), length(sig2)))
)
}
\arguments{
\item{sig1}{vector of first signature}

\item{sig2}{vector of second signature}

\item{min.overlap}{additional parameter passed on to get_ccf}
}
\value{
list consisting of
a) the maximal cross correlation,
b) the lag resulting in the highest cross correlation, and
c) same data frame as input, but y vectors are aligned for
maximal correlation
d) list of a vector of correlations and a vector of lags
}
\description{
Align two surface cross cuts according to maximal correlation
}
