% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_align_set.R
\name{sig_align_set}
\alias{sig_align_set}
\title{Align signatures}
\usage{
sig_align_set(data, value, group)
}
\arguments{
\item{data}{data frame with id, signatures and grouping}

\item{value}{symbol for the signature values}

\item{group}{symbol for grouping}
}
\value{
data frame
}
\description{
Align pairwise across signatures.
}
\examples{
load("data/toolmarks.Rdata")
toolmarks \%>\% group_by(tool, side, size) \%>\% summarize(angles = length(unique(angle)))
tool1AL <- dplyr::filter(toolmarks, tool == "01", side=="A", size=="L")
long <- tool1AL \%>\% group_by(angle) \%>\% nest()
long <- long \%>\% mutate(
  aligned = data \%>\% purrr::map(.f = function(d) {
    d \%>\% sig_align_set(value = signature,  group = mark)
  })
)
long$aligned[[1]] \%>\% ggplot(aes(x =x, y = aligned, colour = factor(mark))) + geom_line()

long <- long \%>\% select(-data) \%>\% unnest(col=aligned)
long \%>\%
  ggplot(aes(x =x, y = aligned, colour = factor(mark))) +
  geom_line() +
  facet_grid(angle~., labeller="label_both")

}
