% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_image.R
\name{x3p_image}
\alias{x3p_image}
\title{Show surface matrix as a plot}
\usage{
x3p_image(x3p, dimension = 3, width = 400, scale = 0.25, texture = "bw")
}
\arguments{
\item{x3p}{scan in x3p format}

\item{dimension}{2 or 3}

\item{width}{size of window}

\item{scale}{factor to scale the original matrix. Large matrices might crash your session.}

\item{texture}{texture object. Create one with \code{rayshader::create_texture} or use one of the pre-made one (e.g. "bw", "imhof1", "imhof2", "imhof3", "unicorn", "desert")}
}
\description{
More details to come
}
