% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p_select_line.R
\name{x3p_select_line}
\alias{x3p_select_line}
\title{Interactively select a line from a scan}
\usage{
x3p_select_line(x3p)
}
\arguments{
\item{x3p}{x3p object containing the scan}
}
\description{
Interactively select a line from a scan
}
