% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_name_pattern.R
\name{assert_name_pattern}
\alias{assert_name_pattern}
\title{Ensure that all files in a vector are correctly named}
\usage{
assert_name_pattern(file, pattern, verbose = FALSE)
}
\arguments{
\item{file}{(vector of) file name(s)}

\item{pattern}{format for a study, generally in regular expression. See e.g. \url{https://regexr.com/} for help in creating suitable patterns}

\item{verbose}{logical value whether function chatters for re-assurance when things works. Defaults to FALSE.}
}
\value{
vector of indices of files that don't pass the check
}
\description{
This function runs a list of file names through a pattern check. A warning
will be returned in case not all file names pass the check. Offending
file names will be shown.
Maria: your specific pattern is defined as
\verb{"^Sc\\\\d\\\\d\\\\.Pl\\\\d\\\\d\\\\d\\\\.Ma\\\\d\\\\.S[AB]\\\\.An\\\\d\\\\d\\\\.Pb\\\\.Dir(Fo|Ba)\\\\.\\\\Siz[SML](|.mid)\\\\.x3p$}
}
\examples{
files <- dir("data-raw", pattern=".x3p", recursive = TRUE, full.names=FALSE)
assert_name_pattern(files, pattern="^T\\\\d\\\\d[SML][AB]-[FB][678]0-0[1-8].x3p$")
}
