% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_align_set.R
\name{sig_align_set}
\alias{sig_align_set}
\title{Align signatures}
\usage{
sig_align_set(data, value, group)
}
\arguments{
\item{data}{data frame with id, signatures and grouping}

\item{value}{symbol for the signature values}

\item{group}{symbol for grouping}
}
\value{
data frame
}
\description{
Align pairwise across signatures.
}
